Conflicts:
	libs/MoSyncLibIOS/MoSyncLibIOS.xcodeproj/project.pbxproj

+++ tools/package/android.cpp
@@ -25,6 +25,8 @@ Software Foundation, 59 Temple Place - Suite 330, Boston, MA
 #include "filelist/filelist.h"
 #include "profiledb/profiledb.h"
 #include <fstream>
+#include <map>
+#include <set>
 #include <sstream>
 #include <errno.h>
 #include <vector>
@@ -42,7 +44,7 @@ static void sign(const SETTINGS& s, const RuntimeInfo& ri, string& unsignedApk, 
 static void createSignCmd(ostringstream& cmd, string& keystore, string& alias, string& storepass, string& keypass, string& signedApk, string& unsignedApk, bool hidden);
 static void writeNFCResource(const SETTINGS& s, const RuntimeInfo& ri);
 static string packageNameToByteCodeName(const string& packageName);
-static string findNativeLibrary(const SETTINGS& s, vector<string>& modules, string& name, string& arch, bool debug);
+static string findNativeLibrary(const SETTINGS& s, vector<string>& modules, string& name, string& arch, bool debug, bool& staticLib);
 
 class AndroidContext : public DefaultContext {
 private:
@@ -99,11 +101,8 @@ void packageAndroid(const SETTINGS& s, const RuntimeInfo& ri) {
 	_mkdir(assets.c_str());
 	if (s.program && existsFile(s.program)) {
 		copyFile(programMp3.c_str(), s.program);
-	} else {
-		// Create dummy file. We'll fix this later
-		// so the runtime ignores it...
-		copyFile(programMp3.c_str(), (string(mosyncdir()) + "/profiles/platforms/Android/dummy.mp3").c_str());
 	}
+
 	if(s.resource && existsFile(s.resource)) {
 		string resMp3 = assets + "/resources.mp3";
 		copyFile(resMp3.c_str(), s.resource);
@@ -134,17 +133,32 @@ void packageAndroid(const SETTINGS& s, const RuntimeInfo& ri) {
 
 	// Extensions & modules.
 	vector<string> modules;
+	set<string> staticModules;
 	map<string,string> initFuncs;
 
 	bool isNative = !strcmp("native", s.outputType);
 
+	// TODO: beta
+	bool useSTL = true;
+	/*for (size_t i = 0; i < extensions.size(); i++) {
+		useSTL |= "stlport" == extensions[i];
+	}*/
 	if (isNative) {
 		// Default modules; todo: externalize?
+		if (useSTL) {
+			modules.push_back("stlport_shared");
+		}
 		modules.push_back("mosync");
 		modules.push_back("mosynclib");
 		initFuncs["mosynclib"] = "resource_selector";
+		//staticModules.insert("mosynclib");
 
-		modules.insert(modules.end(), extensions.begin(), extensions.end());
+		for (size_t i = 0; i < extensions.size(); i++) {
+			string extension = extensions[i];
+			if ("stlport" != extension) {
+				modules.push_back(extension);
+			}
+		}
 		modules.push_back(s.name);
 		initFuncs[string(s.name)] = "MAMain";
 	}
@@ -174,9 +188,12 @@ void packageAndroid(const SETTINGS& s, const RuntimeInfo& ri) {
 		string moduleList = assetDir + "startup.mf";
 		ofstream moduleListOut(moduleList.c_str());
 		for (size_t i = 0; i < modules.size(); i++) {
+			string module = modules[i];
+			moduleListOut << (staticModules.count(module) ?
+					"STATIC:" : "SHARED:");
 			// Last one in list is the 'app' lib
-			moduleListOut << modules[i];
-			string initFunc = initFuncs[modules[i]];
+			moduleListOut << module;
+			string initFunc = initFuncs[module];
 			if (!initFunc.empty()) {
 				moduleListOut << ":" << initFunc;
 			}
@@ -236,13 +253,15 @@ void packageAndroid(const SETTINGS& s, const RuntimeInfo& ri) {
 
 		for (size_t j = 0; j < modules.size(); j++) {
 			string module = modules[j];
-			string nativeLib = findNativeLibrary(s, modules, module, arch, s.debug);
-			if (!nativeLib.empty()) {
+			bool staticLib = false;
+			string nativeLib = findNativeLibrary(s, modules, module, arch, s.debug, staticLib);
+			bool mustExist = !staticLib;//&& "stlport_shared" != module;
+			if (!nativeLib.empty() && mustExist) {
 				string dstLibDir = addlib + "/" + arch + "/";
 				_mkdir(dstLibDir.c_str());
 				string dstLib = dstLibDir + "lib" + module + ".so";
 				copyFile(dstLib.c_str(), nativeLib.c_str());
-			} else {
+			} else if (nativeLib.empty()) {
 				printf("Could not find library %s!\n", module.c_str());
 				exit(1);
 			}
@@ -334,7 +353,7 @@ static void createSignCmd(ostringstream& cmd, string& keystore, string& alias, s
 		" "<<arg(alias);
 }
 
-static string findNativeLibrary(const SETTINGS& s, vector<string>& modules, string& name, string& arch, bool debug) {
+static string findNativeLibrary(const SETTINGS& s, vector<string>& modules, string& name, string& arch, bool debug, bool& staticLib) {
 	vector<string> paths;
 
 	string variantDirName = string("android_") + arch + (debug ? "_debug" : "_release");
@@ -350,17 +369,22 @@ static string findNativeLibrary(const SETTINGS& s, vector<string>& modules, stri
 	// 3. Look in the modules directory
 	for (size_t i = 0; i < modules.size(); i++) {
 		string module = modules[i];
-		string moduleDir = mosyncdir() + string("/modules/") + module + string("/lib/Android/Debug/");
-		string moduleLibPath = moduleDir + arch;
+		string moduleLibPath = mosyncdir() + string("/modules/") + module + string("/lib/") + variantDirName;
 		paths.push_back(moduleLibPath);
 	}
 
 	for (size_t i = 0; i < paths.size(); i++)  {
 		string path = paths[i];
 		toDir(path);
-		string potentialMatch = path + "lib" + name + ".so";
-		if (existsFile(potentialMatch.c_str())) {
-			return potentialMatch;
+		string potentialSharedLib = path + "lib" + name + ".so";
+		string potentialStaticLib = path + "lib" + name + ".a";
+		if (existsFile(potentialSharedLib.c_str())) {
+			staticLib = false;
+			return potentialSharedLib;
+		}
+		if (existsFile(potentialStaticLib.c_str())) {
+			staticLib = true;
+			return potentialStaticLib;
 		}
 	}
 	return "";
